extracted some xml parsing bits from RigidBodyTreeURDF so they can be shared with RigidBodyTreeSDF
